+++ authService/src/entities/User.ts
@@ -14,4 +14,7 @@ export class User extends BaseEntity{
 
     @Column("boolean", { default: false })
     confirmed: boolean;
+
+    @Column("boolean", { default: false })
+    forgotPasswordLocked: boolean;
 }
