+++ authService/src/logout/logout-resolver.ts
@@ -11,18 +11,7 @@ export const logoutResolver:any = {
         logout: async(parent:any, args:any, {redis,session}, info) => {
             const {userId} = session;
             if(userId) {
-                const sessionIds = await redis.lrange(
-                    `${userSessionIdPrefix}${userId}`,
-                    0,
-                    -1
-                  );
-    
-                  const promises = [];
-                  for(let i = 0; i < sessionIds.length; i++) {
-                    promises.push(redis.del(`${redisSessionPrefix}${sessionIds[i]}`))
-                }
-                promises.push(redis.del(`${userSessionIdPrefix}${userId}`))
-                await Promise.all(promises);
+                
                 return true;
             }
